removes styles
I removed the classes with float right and left. First: we should not style IDs.
I created a class l-align--right for the float right.
For the list, I think we should not use floats. I added already l-table and l-table-cell for a table-layout.
In general: we don't style form label, textarea directly. If you wanna add some styles there, then you have to give them a specific class.
If you think about it: you don't wanna have every form with a with of 400px. or every label should not have a width of 90px.